moved file/folder data to separate file. greater_than function written & tested to detect FileObj size some docs written
